Port the kernel to 64-bit (x86-64) long mode
GRUB's Multiboot 1 hand-off only ever promises 32-bit protected mode -
there's no "multiboot but 64-bit" convention - so boot.s's 32-bit
_start now bootstraps itself into long mode before any other kernel
code runs: builds a minimal identity-mapped page table (4 PDPT entries
x 512 2MB pages each, covering the first 4GB), enables PAE, sets the
long-mode bit in the EFER MSR, enables paging, then loads a 64-bit GDT
and far-jumps into _start64.

The 4GB identity map (rather than just the ~1GB the kernel itself
needs) matters: QEMU (and real hardware) commonly maps a VBE linear
framebuffer's PCI BAR near the 4GB boundary (e.g. ~0xFD000000), and
gfx.c's first write to it triple-faulted the machine before this was
mapped.

Also hit and fixed: GCC treats SSE2 as baseline on x86-64 and will
auto-vectorize ordinary loops (memcpy-shaped code especially) into SSE
instructions, which trap as Invalid Opcode unless the kernel has
explicitly enabled SSE (CR0/CR4 bits) and set up FPU/SSE state
save/restore across interrupts - neither of which this kernel does, on
purpose, since it has no real floating-point needs. Building with
-mgeneral-regs-only stops the compiler from ever emitting them.

Rewrote the ISR/IRQ assembly stubs for the 64-bit register set and
iretq, struct registers in kernel.h to match, and gdt.c/idt.c for the
64-bit descriptor formats (16-byte IDT gates now carry a full 64-bit
offset). linker.ld and the Makefile now target elf64-x86-64 throughout.

Verified end-to-end in qemu-system-x86_64: boot log, GUI (windows,
wallpapers, search, RTC clock), mouse/keyboard input, and the
GUI<->shell round trip via Ctrl+Q all work identically to the 32-bit
build.

+++ kernel/memory.c
@@ -43,7 +43,7 @@ struct mmap_entry {
 static struct multiboot_info *mbi = 0;
 
 void memory_init(uint32_t mb_info_addr) {
-	mbi = (struct multiboot_info *)mb_info_addr;
+	mbi = (struct multiboot_info *)(uintptr_t)mb_info_addr;
 }
 
 uint32_t memory_total_kb(void) {
@@ -97,12 +97,12 @@ void memory_print_map(void) {
 
 	if (mbi->flags & (1 << 6)) {
 		kprintf("\nMemory map:\n");
-		struct mmap_entry *entry = (struct mmap_entry *)mbi->mmap_addr;
-		uint32_t end = mbi->mmap_addr + mbi->mmap_length;
-		while ((uint32_t)entry < end) {
+		struct mmap_entry *entry = (struct mmap_entry *)(uintptr_t)mbi->mmap_addr;
+		uintptr_t end = (uintptr_t)mbi->mmap_addr + mbi->mmap_length;
+		while ((uintptr_t)entry < end) {
 			kprintf("  addr=0x%x len=0x%x type=%d\n",
 				(uint32_t)entry->addr, (uint32_t)entry->len, entry->type);
-			entry = (struct mmap_entry *)((uint32_t)entry + entry->size + 4);
+			entry = (struct mmap_entry *)((uintptr_t)entry + entry->size + 4);
 		}
 	}
 }
